+++ src/main/java/com/bc/service/AddressService.java
@@ -10,11 +10,11 @@ import com.bc.model.Address;
 
 public interface AddressService {
 
-	public Address updateAddressByUserId(Address address, Integer userId, String key)
+	public Address updateAddressByUserId(Address address, Long userId, String key)
 			throws AddressException, CustomerException, SessionLoginException;
 
 	public List<Address> viewAllAddress(String key) throws AddressException, SessionLoginException;
 
-	public Address viewAddressByUserId(Integer userId, String key) throws CustomerException, SessionLoginException;
+	public Address viewAddressByUserId(Long userId, String key) throws CustomerException, SessionLoginException;
 
 }
