+++ src/main/java/com/bc/service/CustomerService.java
@@ -11,7 +11,7 @@ public interface CustomerService {
 
 	public Customer updateCustomer(Customer customer) throws CustomerException;
 
-	public Customer remove(Integer customerId) throws CustomerException;
+	public Customer remove(Long customerId) throws CustomerException;
 
 	public List<Customer> viewAllCustomer() throws CustomerException;
